Screenshots implementation kraken
